Fix cost calculation modal wiring and localization
