+++ Sources/Features/PricesFeature/PricesView.swift
@@ -1,9 +1,6 @@
 import SwiftUI
 
 struct PricesView: View {
-    let onCalculationDurationHoursChanged: (Double) -> Void
-    let onCalculationPlaceholderDismissed: () -> Void
-    let onCalculationPresetSelected: (AppliancePreset.Kind) -> Void
     let onHourTapped: (HourlyPrice) -> Void
     let state: PricesFeature.State
 
@@ -32,19 +29,6 @@ struct PricesView: View {
         .background(Color(.systemBackground))
         .accessibilityIdentifier("pricesScreen")
         .safeAreaPadding(.top, PricesViewLayout.safeAreaTopPadding)
-        .sheet(isPresented: isCalculationSheetPresentedBinding) {
-            PricesCalculationSheetView(
-                durationHours: state.costCalculation.durationHours,
-                durationHoursBinding: calculationDurationBinding,
-                onCloseTapped: onCalculationPlaceholderDismissed,
-                presetBinding: selectedPresetBinding,
-                selectedHour: state.costCalculation.selectedHour,
-                selectedPreset: PricesPresetCatalog.preset(for: state.costCalculation.selectedPresetKind),
-                presets: PricesPresetCatalog.all
-            )
-            .presentationDetents([.medium])
-            .presentationDragIndicator(.visible)
-        }
     }
 
     private var cacheBadge: some View {
@@ -57,24 +41,6 @@ struct PricesView: View {
         .accessibilityIdentifier("pricesCacheBadge")
     }
 
-    private var calculationDurationBinding: Binding<Double> {
-        Binding(
-            get: { state.costCalculation.durationHours },
-            set: { onCalculationDurationHoursChanged($0) }
-        )
-    }
-
-    private var isCalculationSheetPresentedBinding: Binding<Bool> {
-        Binding(
-            get: { state.costCalculation.isPresented },
-            set: { isPresented in
-                if !isPresented {
-                    onCalculationPlaceholderDismissed()
-                }
-            }
-        )
-    }
-
     private var noSummaryView: some View {
         Text(String(localized: "prices.summary.empty"))
             .font(.subheadline)
@@ -82,12 +48,6 @@ struct PricesView: View {
             .accessibilityIdentifier("pricesSummaryEmpty")
     }
 
-    private var selectedPresetBinding: Binding<AppliancePreset.Kind> {
-        Binding(
-            get: { state.costCalculation.selectedPresetKind },
-            set: { onCalculationPresetSelected($0) }
-        )
-    }
 }
 
 enum PricesViewLayout {
@@ -108,116 +68,6 @@ enum PricesViewLayout {
     static let sectionSpacing: CGFloat = 18
     static let summaryCardSpacing: CGFloat = 6
     static let verticalSpacing: CGFloat = 16
-}
-
-private enum PricesPresetCatalog {
-    private static let airConditionerPowerKW = 1.2
-    private static let clothesDryerPowerKW = 2.5
-    private static let dishwasherPowerKW = 1.5
-    private static let electricHeaterPowerKW = 1.8
-    private static let electricVehiclePowerKW = 7.2
-    private static let inductionCooktopPowerKW = 1.8
-    private static let ovenPowerKW = 2.2
-    private static let washingMachinePowerKW = 2.0
-    private static let waterHeaterPowerKW = 2.0
-
-    static var all: [AppliancePreset] {
-        [
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.airConditioner.description"),
-                kind: .airConditioner,
-                displayName: String(localized: "prices.calculation.preset.airConditioner.name"),
-                powerKW: airConditionerPowerKW,
-                symbolName: "fan"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.clothesDryer.description"),
-                kind: .clothesDryer,
-                displayName: String(localized: "prices.calculation.preset.clothesDryer.name"),
-                powerKW: clothesDryerPowerKW,
-                symbolName: "wind"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.dishwasher.description"),
-                kind: .dishwasher,
-                displayName: String(localized: "prices.calculation.preset.dishwasher.name"),
-                powerKW: dishwasherPowerKW,
-                symbolName: "drop"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.electricHeater.description"),
-                kind: .electricHeater,
-                displayName: String(localized: "prices.calculation.preset.electricHeater.name"),
-                powerKW: electricHeaterPowerKW,
-                symbolName: "thermometer.medium"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.electricVehicle.description"),
-                kind: .electricVehicle,
-                displayName: String(localized: "prices.calculation.preset.electricVehicle.name"),
-                powerKW: electricVehiclePowerKW,
-                symbolName: "car"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.inductionCooktop.description"),
-                kind: .inductionCooktop,
-                displayName: String(localized: "prices.calculation.preset.inductionCooktop.name"),
-                powerKW: inductionCooktopPowerKW,
-                symbolName: "cooktop"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.oven.description"),
-                kind: .oven,
-                displayName: String(localized: "prices.calculation.preset.oven.name"),
-                powerKW: ovenPowerKW,
-                symbolName: "flame"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.washingMachine.description"),
-                kind: .washingMachine,
-                displayName: String(localized: "prices.calculation.preset.washingMachine.name"),
-                powerKW: washingMachinePowerKW,
-                symbolName: "washer"
-            ),
-            preset(
-                shortDescription: String(localized: "prices.calculation.preset.waterHeater.description"),
-                kind: .waterHeater,
-                displayName: String(localized: "prices.calculation.preset.waterHeater.name"),
-                powerKW: waterHeaterPowerKW,
-                symbolName: "drop.degreesign"
-            ),
-        ]
-    }
-
-    static func preset(for kind: AppliancePreset.Kind) -> AppliancePreset {
-        all.first(where: { $0.kind == kind }) ?? fallbackPreset
-    }
-
-    private static var fallbackPreset: AppliancePreset {
-        preset(
-            shortDescription: String(localized: "prices.calculation.preset.washingMachine.description"),
-            kind: .washingMachine,
-            displayName: String(localized: "prices.calculation.preset.washingMachine.name"),
-            powerKW: washingMachinePowerKW,
-            symbolName: "washer"
-        )
-    }
-
-    private static func preset(
-        shortDescription: String,
-        kind: AppliancePreset.Kind,
-        displayName: String,
-        powerKW: Double,
-        symbolName: String
-    ) -> AppliancePreset {
-        AppliancePreset(
-            displayName: displayName,
-            kind: kind,
-            powerKW: powerKW,
-            shortDescription: shortDescription,
-            symbolName: symbolName
-        )
-    }
 }
 
 enum PricesViewFormatting {
